Split word search into its own file too.
